Enable telemetry, fix channel mappings
Probably the biggest change. BetaFPV source used Bayang BLE with no telemetry enabled, and only 10 channels. I substituted this with the source from rx_bayang_protocol_telemetry_autobind.c (keeping the switching capabilities), enabling channels 11 and 12. 
Also fixed the channel mappings when using "T8SG" setting (this is used in place of the transmitter selection in config.h). Specifically channel 5 and 6 were swapped.
LEDS_ON channel was not working, because it is overridden by the OSD setting - i made it so that if OSD setting is off, then the channel can be used to switch LEDS on/off (when the OSD LED setting is ON, then this overrides the channel setting)