Fix channel mappings, add DFU file creation tools
Channel mappings were incorrect for T8SG (Deviation) controllers.
Also channels 11 and 12 were missing, and no support for analog aux
LED channel was overridden by OSD value so could not be switched via transmitter

Added some tools to automatically generate a DFU file on build. This also attempts to write the DFU file to the FC but will just skip that if the FC is not connected